add feature manager restaurants

+++ src/main/java/com/manager/restaurant/service/RestaurantService.java
@@ -3,14 +3,14 @@ package com.manager.restaurant.service;
 import com.manager.restaurant.dto.request.RestaurantRequest;
 import com.manager.restaurant.dto.request.UpdateRestaurantRequest;
 import com.manager.restaurant.dto.response.RestaurantResponse;
-import com.manager.restaurant.entity.Account;
-import com.manager.restaurant.entity.AccountRole;
-import com.manager.restaurant.entity.Restaurant;
+import com.manager.restaurant.entity.*;
 import com.manager.restaurant.exception.BadException;
 import com.manager.restaurant.exception.ErrorCode;
 import com.manager.restaurant.mapper.RestaurantMapper;
 import com.manager.restaurant.repository.AccountRepository;
 import com.manager.restaurant.repository.RestaurantRepository;
+import com.manager.restaurant.repository.RestaurantsOfHostRepository;
+import com.manager.restaurant.repository.TableRepository;
 import com.manager.restaurant.until.SecurityUtils;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
@@ -18,6 +18,7 @@ import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -28,6 +29,8 @@ public class RestaurantService {
     RestaurantRepository restaurantRepository;
     AccountRepository accountRepository;
     RestaurantMapper restaurantMapper;
+    RestaurantsOfHostRepository restaurantsOfHostRepository;
+    TableRepository tableRepository;
 
     public void createRestaurant(RestaurantRequest request) {
         Account account = accountRepository.findByUsername(SecurityUtils.getCurrentUsername()).orElseThrow(
@@ -36,16 +39,24 @@ public class RestaurantService {
         Restaurant restaurant = restaurantMapper.toRestaurant(request);
         restaurant.setStatus("active");
         restaurantRepository.save(restaurant);
-        account.setRestaurant(restaurant);
-        accountRepository.save(account);
+        restaurantsOfHostRepository.save(new RestaurantsOfHost(restaurant.getIdRestaurant(), account.getIdAccount()));
     }
 
 
-    public void updateNameRestaurant(UpdateRestaurantRequest request) {
+    public void updateRestaurant(UpdateRestaurantRequest request) {
         Restaurant restaurant = restaurantRepository.findById(request.getIdRestaurant()).orElseThrow(
                 ()-> new BadException(ErrorCode.RESTAURANT_NOT_FOUND)
         );
         restaurant.setName(request.getName());
+        if(!request.getStatus().equals("Active")){
+            restaurant.setStatus("Inactive");
+            List<Account> accounts = accountRepository.findAllByRestaurant_IdRestaurant(restaurant.getIdRestaurant());
+            accounts.removeIf(acc->acc.getRole().equals(AccountRole.Owner.toString()));
+            for (Account ac: accounts) {
+                ac.setStatus("Inactive");
+            }
+            accountRepository.saveAll(accounts);
+        }
         restaurantRepository.save(restaurant);
     }
 
@@ -57,9 +68,21 @@ public class RestaurantService {
             Restaurant restaurant = restaurantRepository.findById(idRestaurant).orElseThrow(
                     ()-> new BadException(ErrorCode.RESTAURANT_NOT_FOUND)
             );
+            if(tableRepository.existsByRestaurantAndStatus(restaurant,"Unavailable")){
+                throw new BadException(ErrorCode.CANT_DELETE_RES);
+            }
+
+            RestaurantsOfHostPK restaurantsOfHostPK = new RestaurantsOfHostPK(account.getIdAccount(), restaurant.getIdRestaurant());
+            if(!restaurantsOfHostRepository.existsById(restaurantsOfHostPK)){
+                throw new BadException(ErrorCode.ACCESS_DENIED);
+            }
+            restaurantsOfHostRepository.deleteById(restaurantsOfHostPK);
             List<Account> accounts = accountRepository.findAllByRestaurant_IdRestaurant(idRestaurant);
             accounts.removeIf(acc->acc.getRole().equals(AccountRole.Owner.toString()));
-            accountRepository.deleteAll(accounts);
+            for (Account ac: accounts) {
+                ac.setStatus("Inactive");
+            }
+            accountRepository.saveAll(accounts);
             restaurantRepository.delete(restaurant);
             account.setRestaurant(null);
             accountRepository.save(account);
@@ -74,10 +97,19 @@ public class RestaurantService {
         ));
     }
 
-    public RestaurantResponse getMyRestaurant() {
+    public List<RestaurantResponse> getMyRestaurant() {
         Account account = accountRepository.findByUsername(SecurityUtils.getCurrentUsername()).orElseThrow(
                 ()-> new BadException(ErrorCode.USER_NOT_EXISTED)
         );
-        return restaurantMapper.toRestaurantResponse(account.getRestaurant());
+        List<RestaurantResponse> rp = new ArrayList<>();
+        if(AccountRole.Owner.toString().equals(account.getRole())){
+            List<RestaurantsOfHost> restaurantsOfHosts = restaurantsOfHostRepository.findByIdAccount(account.getIdAccount());
+            for (RestaurantsOfHost resOfHost: restaurantsOfHosts) {
+                rp.add(restaurantMapper.toRestaurantResponse(restaurantRepository.findById(resOfHost.getIdRestaurant()).orElseThrow()));
+            }
+        }else {
+            rp.add(restaurantMapper.toRestaurantResponse(account.getRestaurant()));
+        }
+        return rp;
     }
 }
